requirements.txt
flask
pytesseract
pillow
transformers
torch

openai
python-dotenv